// LMS_Frontend/src/app/models/sendUser.model.ts
//  FOR NORMAL USE
export interface sendUser {
	userName: string;
	userEmail: string;
	userPhoneNumber: number;
	userGovId: string;
	userPassword: string;
	userAddress: string;
}


// FORM using iTURMERIC
// export interface sendUser {
// 	userId : string;
// 	userName: string;
// 	userEmail: string;
// 	userPhoneNumber: number;
// 	userGovId: string;
// 	userPassword: string;
// 	userAddress: string;
// }